requirements.txt
opencv-python==4.7.0.72
easyocr
Pillow
matplotlib
plotly
pandas
numpy
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector
python-multipart
